requirements.txt
flask==3.0.0
python-pptx==0.6.23
gunicorn==21.2.0
openpyxl==3.1.2
werkzeug==3.0.1
